Eliminado nullcheck raruno
No es muy best practices lanzar una NullPointerException y capturarla
para ignorarla cuando se resuelve con un modesto if. Asín mejor.

+++ domain/src/main/java/com/shootr/android/domain/interactor/stream/RemoveFromFavoritesInteractor.java
@@ -40,13 +40,10 @@ public class RemoveFromFavoritesInteractor implements Interactor {
 
     @Override public void execute() throws Exception {
         Favorite existingFavorite = localFavoriteRepository.getFavoriteByStream(idStream);
-        try {
-            checkNotNull(existingFavorite);
+        if (existingFavorite != null) {
             localFavoriteRepository.removeFavoriteByStream(existingFavorite.getIdStream());
             notifyCompleted();
             remoteFavoriteRepository.removeFavoriteByStream(existingFavorite.getIdStream());
-        } catch (NullPointerException error) {
-            // DO nothing
         }
     }
